tests/wpt/web-platform-tests/css/css-scoping/keyframes-005.html

<!doctype html>
<title>CSS Test: @keyframes should not leak out of shadow tree.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://drafts.csswg.org/css-scoping/#shadow-names">
<style>
  #host {
    width: 100px;
    height: 100px;
    background: blue;
    animation: myanim 10s infinite;
  }
</style>
<div id="host"></div>
<script>
  test(function() {
    host.attachShadow({ mode: "open" }).innerHTML = `
      <style>
        @keyframes myanim {
          from { background: red; }
          to { background: green; }
        }
      </style>
    `;

    assert_equals(document.getElementById('host').getAnimations().length, 0);
  }, "@keyframes should not leak out of the shadow tree.");
</script>
